Use the last modified date of a page to check if it changed
Previously we would request the page, generate an md5 of the body and
check if it was the same saved in the database.
Because of this we could send an image using the file id returned by the
telegram api, making the bot faster and removing the image download time
if the user already checked that page and it didn't change since then.
Now we are doing a conditional GET request with the If-Modified-Since
header set with the last time we checked if the page changed.
Other than making this faster and removing a column from the database
(the column used to save the md5 of the page body), but it also bypasses
the cache if the image with the real table changes (hoping they continue
to deploy both the html file and the jpg file when a page changes)